chore: move frontpage programmes out of accordion
