+++ e2e/specs/unauthenticated/frontpage.spec.ts
@@ -15,7 +15,6 @@ test.beforeEach(async ({ page }) => {
 
 test("should have list of valid links on frontpage", async ({ page }) => {
   await mockWaitResponse(page, "**/graphql-api/*");
-  await page.getByTestId("accordion-header").click();
   const programmes = page.getByTestId("programme-list").getByRole("link");
   await expect(programmes).toHaveCount(16);
 });
