+++ package.json
@@ -43,7 +43,7 @@
     "@graphql-codegen/typescript": "^4.0.6",
     "@graphql-codegen/typescript-operations": "^4.2.0",
     "@graphql-eslint/eslint-plugin": "^3.20.1",
-    "@ndla/preset-panda": "^0.0.27",
+    "@ndla/preset-panda": "^0.0.28",
     "@ndla/scripts": "^2.1.2",
     "@ndla/types-backend": "^0.2.88",
     "@ndla/types-embed": "^5.0.1-alpha.0",
@@ -91,18 +91,18 @@
     "@fontsource/source-sans-pro": "^4.5.9",
     "@fontsource/source-serif-pro": "^4.5.7",
     "@lexical/react": "^0.12.4",
-    "@ndla/article-converter": "^10.0.26-alpha.0",
+    "@ndla/article-converter": "^10.0.27-alpha.0",
     "@ndla/core": "^5.0.2",
     "@ndla/error-reporter": "^2.0.2",
     "@ndla/hooks": "^2.1.9",
-    "@ndla/icons": "^8.0.17-alpha.0",
-    "@ndla/licenses": "^8.0.2-alpha.0",
-    "@ndla/primitives": "^1.0.23-alpha.0",
-    "@ndla/safelink": "^7.0.23-alpha.0",
-    "@ndla/styled-system": "^0.0.19",
+    "@ndla/icons": "^8.0.18-alpha.0",
+    "@ndla/licenses": "^8.0.3-alpha.0",
+    "@ndla/primitives": "^1.0.24-alpha.0",
+    "@ndla/safelink": "^7.0.24-alpha.0",
+    "@ndla/styled-system": "^0.0.20",
     "@ndla/tracker": "^5.0.7",
     "@ndla/typography": "^0.4.24",
-    "@ndla/ui": "^56.0.26-alpha.0",
+    "@ndla/ui": "^56.0.27-alpha.0",
     "@ndla/util": "^4.1.0",
     "cross-env": "^7.0.3",
     "date-fns": "^2.30.0",
